# Practice Projects/MicroService/CommercialService/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Commercial Service API",
    "description": "API documentation for Commercial Service - managing owners, properties, and payments",
    "version": "1.0.0",
    "contact": {
      "name": "Commercial Service Team"
    },
    "license": {
      "name": "Apache 2.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8090",
      "description": "Development Server"
    },
    {
      "url": "http://api.commercial-service.com",
      "description": "Production Server"
    }
  ],
  "tags": [
    {
      "name": "Owner",
      "description": "APIs for managing property owners"
    },
    {
      "name": "Property",
      "description": "APIs for managing properties"
    },
    {
      "name": "Payment",
      "description": "APIs for managing payments"
    }
  ],
  "paths": {
    "/api/v1/owner/register": {
      "post": {
        "summary": "Register Owner",
        "description": "API to register a new owner in the system",
        "operationId": "registerOwner",
        "tags": [
          "Owner"
        ],
        "requestBody": {
          "required": true,
          "description": "Owner registration details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnerRequestDTO"
              },
              "examples": {
                "example1": {
                  "summary": "Basic owner registration",
                  "value": {
                    "name": "John Doe",
                    "email": "john.doe@example.com",
                    "phone": "+1-555-1234",
                    "ownerAddressDetails": [
                      {
                        "addressType": "Residential",
                        "addressLine1": "123 Main Street",
                        "addressLine2": "Apt 4B",
                        "city": "New York",
                        "state": "NY",
                        "country": "USA"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Owner registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerResponseDTO"
                },
                "examples": {
                  "example1": {
                    "summary": "Successful registration response",
                    "value": {
                      "id": 1,
                      "name": "John Doe",
                      "email": "john.doe@example.com",
                      "phone": "+1-555-1234",
                      "ownerAddressDetails": [
                        {
                          "addressType": "Residential",
                          "addressLine1": "123 Main Street",
                          "addressLine2": "Apt 4B",
                          "city": "New York",
                          "state": "NY",
                          "country": "USA"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Owner already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/owner/{ownerId}": {
      "get": {
        "summary": "Get Owner by ID",
        "description": "API to retrieve owner details by their ID",
        "operationId": "getOwnerById",
        "tags": [
          "Owner"
        ],
        "parameters": [
          {
            "name": "ownerId",
            "in": "path",
            "description": "ID of the owner to retrieve",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Owner details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Owner not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/property/getPropertyByOwnerId": {
      "get": {
        "summary": "Get Properties of Owner",
        "description": "Retrieve all properties associated with a specific owner",
        "operationId": "getPropertyByOwnerId",
        "tags": [
          "Property"
        ],
        "parameters": [
          {
            "name": "ownerId",
            "in": "query",
            "description": "ID of the owner to retrieve properties for",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Properties retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PropertyDTO"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Owner not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/payAdvance": {
      "post": {
        "summary": "Pay Advance for Property",
        "description": "Process advance payment for a property",
        "operationId": "payAdvanceForProperty",
        "tags": [
          "Payment"
        ],
        "requestBody": {
          "required": true,
          "description": "Payment details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentRequestDTO"
              },
              "examples": {
                "example1": {
                  "summary": "Sample payment request",
                  "value": {
                    "payment": {
                      "id": 1,
                      "propertyId": 1,
                      "ownerId": 1,
                      "amount": 50000,
                      "paymentDate": "2026-01-30",
                      "transactionId": "TXN123456",
                      "status": "Pending"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Advance payment processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponseDTO"
                },
                "examples": {
                  "example1": {
                    "summary": "Successful payment response",
                    "value": {
                      "amount": 50000,
                      "transactionId": "TXN123456",
                      "message": "Payment processed successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid payment details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "OwnerRequestDTO": {
        "type": "object",
        "description": "Owner registration request data",
        "required": [
          "name",
          "email",
          "phone"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the owner",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the owner",
            "example": "john.doe@example.com"
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the owner",
            "example": "+1-555-1234"
          },
          "ownerAddressDetails": {
            "type": "array",
            "description": "List of addresses associated with the owner",
            "items": {
              "$ref": "#/components/schemas/OwnerAddressRequestDTO"
            }
          }
        }
      },
      "OwnerAddressRequestDTO": {
        "type": "object",
        "description": "Owner address information for registration",
        "required": [
          "addressType",
          "addressLine1",
          "city",
          "state",
          "country"
        ],
        "properties": {
          "addressType": {
            "type": "string",
            "description": "Type of address (e.g., Residential, Commercial, Office)",
            "example": "Residential"
          },
          "addressLine1": {
            "type": "string",
            "description": "Primary address line",
            "example": "123 Main Street"
          },
          "addressLine2": {
            "type": "string",
            "description": "Secondary address line (optional)",
            "example": "Apt 4B"
          },
          "city": {
            "type": "string",
            "description": "City name",
            "example": "New York"
          },
          "state": {
            "type": "string",
            "description": "State or province name",
            "example": "NY"
          },
          "country": {
            "type": "string",
            "description": "Country name",
            "example": "USA"
          }
        }
      },
      "OwnerResponseDTO": {
        "type": "object",
        "description": "Owner response data",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique identifier of the owner",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Full name of the owner",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the owner",
            "example": "john.doe@example.com"
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the owner",
            "example": "+1-555-1234"
          },
          "ownerAddressDetails": {
            "type": "array",
            "description": "List of addresses associated with the owner",
            "items": {
              "$ref": "#/components/schemas/OwnerAddressResponseDTO"
            }
          }
        }
      },
      "OwnerAddressResponseDTO": {
        "type": "object",
        "description": "Owner address information in response",
        "properties": {
          "addressType": {
            "type": "string",
            "description": "Type of address",
            "example": "Residential"
          },
          "addressLine1": {
            "type": "string",
            "description": "Primary address line",
            "example": "123 Main Street"
          },
          "addressLine2": {
            "type": "string",
            "description": "Secondary address line",
            "example": "Apt 4B"
          },
          "city": {
            "type": "string",
            "description": "City name",
            "example": "New York"
          },
          "state": {
            "type": "string",
            "description": "State or province name",
            "example": "NY"
          },
          "country": {
            "type": "string",
            "description": "Country name",
            "example": "USA"
          }
        }
      },
      "PropertyDTO": {
        "type": "object",
        "description": "Property information",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique identifier of the property",
            "example": 1
          },
          "title": {
            "type": "string",
            "description": "Title of the property",
            "example": "Modern Office Space"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the property",
            "example": "A modern office space with state-of-the-art facilities"
          },
          "type": {
            "type": "string",
            "description": "Type of property (Office, Warehouse, Retail, etc.)",
            "enum": [
              "Office",
              "Warehouse",
              "Retail",
              "Commercial"
            ],
            "example": "Office"
          },
          "location": {
            "type": "string",
            "description": "Location/address of the property",
            "example": "123 Business Ave, New York, NY"
          },
          "price": {
            "type": "number",
            "format": "double",
            "description": "Price of the property",
            "example": 500000
          },
          "status": {
            "type": "string",
            "description": "Current status of the property",
            "enum": [
              "Available",
              "Sold",
              "Rented",
              "Under Maintenance"
            ],
            "example": "Available"
          },
          "advancePayment": {
            "type": "number",
            "format": "double",
            "description": "Advance payment amount",
            "example": 50000
          },
          "ownerId": {
            "type": "integer",
            "format": "int64",
            "description": "ID of the owner",
            "example": 1
          }
        }
      },
      "PaymentRequestDTO": {
        "type": "object",
        "description": "Payment request data",
        "properties": {
          "payment": {
            "$ref": "#/components/schemas/Payment"
          }
        }
      },
      "Payment": {
        "type": "object",
        "description": "Payment details",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique identifier of the payment",
            "example": 1
          },
          "propertyId": {
            "type": "integer",
            "format": "int64",
            "description": "ID of the property",
            "example": 1
          },
          "ownerId": {
            "type": "integer",
            "format": "int64",
            "description": "ID of the owner",
            "example": 1
          },
          "amount": {
            "type": "number",
            "format": "double",
            "description": "Payment amount",
            "example": 50000
          },
          "paymentDate": {
            "type": "string",
            "format": "date",
            "description": "Date of payment",
            "example": "2026-01-30"
          },
          "transactionId": {
            "type": "string",
            "description": "Unique transaction identifier",
            "example": "TXN123456"
          },
          "status": {
            "type": "string",
            "description": "Payment status",
            "enum": [
              "Pending",
              "Completed",
              "Failed",
              "Cancelled"
            ],
            "example": "Pending"
          }
        }
      },
      "PaymentResponseDTO": {
        "type": "object",
        "description": "Payment response data",
        "properties": {
          "amount": {
            "type": "number",
            "format": "double",
            "description": "Payment amount",
            "example": 50000
          },
          "transactionId": {
            "type": "string",
            "description": "Unique transaction identifier",
            "example": "TXN123456"
          },
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Payment processed successfully"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Error response structure",
        "properties": {
          "status": {
            "type": "integer",
            "description": "HTTP status code",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "Invalid input data"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of the error",
            "example": "2026-01-30T10:30:00Z"
          },
          "path": {
            "type": "string",
            "description": "API endpoint that caused the error",
            "example": "/api/v1/owner/register"
          }
        }
      }
    }
  }
}