# Practice Projects/MicroService/CommercialService/src/main/java/com/org/CommercialService/Controller/OwnerController.java
package com.org.CommercialService.Controller;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import com.org.CommercialService.DTO.Request.OwnerRequestDTO;
import com.org.CommercialService.DTO.Response.OwnerResponseDTO;
import com.org.CommercialService.Service.OwnerService;
import com.org.CommercialService.Service.PropertyService;
import com.org.CommercialService.swagger.OwnerAPI;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;

@RestController
@RequestMapping("api/v1/owner")
public class OwnerController implements OwnerAPI{

    private final OwnerService ownerService;
    private final PropertyService propertyService;

    @Autowired
    public OwnerController(OwnerService ownerService, PropertyService propertyService) {
        this.ownerService = ownerService;
        this.propertyService = propertyService;
    }

    
    @Override
    public ResponseEntity<OwnerResponseDTO> registerOwner(@Valid @RequestBody OwnerRequestDTO ownerRequestDTO) {
        System.out.println("ownerRequestDTO >> " + ownerRequestDTO);
        OwnerResponseDTO responseDTO = ownerService.registerOwner(ownerRequestDTO);
        return ResponseEntity.ok(responseDTO);
    }

}